requirements.txt
anthropic==0.84.0
python-dotenv==1.2.2
